lasr: Move everything into a lasr/ subdirectory
