+++ meson.build
@@ -28,16 +28,27 @@ executable(
     'libresplit',
     [
         'src/main.c',
-        'src/shared.c',
-        'src/server.c',
-        'src/auto-splitter.c',
         'src/bind.c',
-        'src/memory.c',
-        'src/process.c',
+        'src/server.c',
         'src/settings.c',
-        'src/signature.c',
+        'src/shared.c',
         'src/timer.c',
-        'src/lua_utils.c',
+
+        # LASR
+        'src/lasr/auto-splitter.c',
+        'src/lasr/utils.c',
+        'src/lasr/functions/bitwise.c',
+        'src/lasr/functions/getBaseAddress.c',
+        'src/lasr/functions/getModuleSize.c',
+        'src/lasr/functions/getPID.c',
+        'src/lasr/functions/print_tbl.c',
+        'src/lasr/functions/process.c',
+        'src/lasr/functions/readAddress.c',
+        'src/lasr/functions/shallow_copy_tbl.c',
+        'src/lasr/functions/signature.c',
+        'src/lasr/functions/sizeOf.c',
+
+        # Components
         'src/component/best-sum.c',
         'src/component/clock.c',
         'src/component/components.c',
