+++ src/lasr/functions.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "functions/bitwise.h"
+#include "functions/getBaseAddress.h"
+#include "functions/getModuleSize.h"
+#include "functions/getPID.h"
+#include "functions/print_tbl.h"
+#include "functions/process.h"
+#include "functions/readAddress.h"
+#include "functions/shallow_copy_tbl.h"
+#include "functions/signature.h"
+#include "functions/sizeOf.h"
