+++ src/lasr/functions/getBaseAddress.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <lua.h>
+
+int getBaseAddress(lua_State* L);
