+++ src/lasr/functions/getModuleSize.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <lua.h>
+
+int getModuleSize(lua_State* L);
