+++ src/lasr/functions/print_tbl.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <lua.h>
+
+int print_tbl(lua_State* L);
