+++ src/lasr/functions/process.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <lua.h>
+
+int find_process_id(lua_State* L);
