+++ src/lasr/functions/shallow_copy_tbl.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <lua.h>
+
+int shallow_copy_tbl(lua_State* L);
