+++ src/lasr/functions/signature.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <lua.h>
+
+int perform_sig_scan(lua_State* L);
