+++ src/lasr/functions/sizeOf.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <lua.h>
+
+int size_of(lua_State* L);
