--- src/process.h
@@ -1,36 +1,0 @@
-#ifndef __PROCESS_H__
-#define __PROCESS_H__
-
-#include <linux/limits.h>
-#include <stdbool.h>
-#include <stdint.h>
-
-#include <luajit.h>
-
-/**
- * \struct game_process The game process read by the Auto Splitter
- */
-struct game_process {
-    const char* name; /*!< The name of the process */
-    int pid; /*!< The PID of the process */
-    uintptr_t base_address; /*!< The detected base address of the process */
-    uintptr_t dll_address; /*!< The detected base address of the last requested module */
-};
-typedef struct game_process game_process;
-
-typedef struct ProcessMap {
-    uintptr_t start;
-    uintptr_t end;
-    uintptr_t size;
-    char name[PATH_MAX];
-} ProcessMap;
-extern uint32_t p_maps_cache_size;
-
-uintptr_t find_base_address(const char* module);
-int process_exists();
-int find_process_id(lua_State* L);
-int getPid(lua_State* L);
-bool parseMapsLine(const char* line, ProcessMap* map);
-int lua_get_module_size(lua_State* L);
-
-#endif /* __PROCESS_H__ */
